Escritura del nuevo producto implementado (sin embargo hay errores que provocan un crasheo del sistema)

+++ WNS/src/wns/Product.java
@@ -143,7 +143,7 @@ public class Product {
      * Price cannot be a negative or zero number.
      * @param newPrice
      */
-    public void setPrice (int newPrice) {
+    public void setPrice (double newPrice) {
         if (newPrice <= 0)
             System.out.println("\tPrice should be nonnegative!");
         else
